Compare highlight count with sidebar tab counts during testing
Previously the testing tool just waited for the sidebar to display
counts next to its tabs. However this does not account for the fact that
the tab counts can change during the anchoring process as annotations
initially show up optimistically in the "Annotations" tab but may later
transfer to the "Orphans" tab if they are not found in the page.

There is also a possibility of bugs where an annotation is counted in
the Annotations tab but there is no highlight in the page.

The solution here is to count the number of highlighted annotations
(that is, the number of annotation tags with an associated highlight)
and compare that to the tab counts. Anchoring is assumed to have
completed when the highlight count matches the annotation count.