Limit anchoring timeout once at least one highlight has appeared in the page
There seems to be a common issue at the moment with PDFs where the total
number of annotations with at least one highlight does not match the
annotation count in the sidebar. In this case there appear to be
annotations which are orphaned in that clicking them in the sidebar will
not scroll to a location in the document, but the annotation still
appears in the annotations tab in the page.

A possible cause of this is known issue with rendering of overlapping
highlights.